Use Safir's spawn_uvicorn for monkeyflocker testing
We have to spawn an external application for monkeyflocker testing.
Use Safir's spawn_uvicorn rather than rolling our own, and use the
factory approach so that we can get proper type checking and black
formatting of the app.
